refactor(wallet-mobile): feature flags
