+++ apps/wallet-mobile/src/components/ConfirmTx/ConfirmTx.tsx
@@ -6,10 +6,10 @@ import React, {useEffect, useState} from 'react'
 import {useIntl} from 'react-intl'
 import {StyleSheet, View} from 'react-native'
 
-import {debugWalletInfo, features} from '../../features'
 import {useAuthOsWithEasyConfirmation} from '../../features/Auth/common/hooks'
 import {useSelectedWallet} from '../../features/WalletManager/common/hooks/useSelectedWallet'
 import {useWalletManager} from '../../features/WalletManager/context/WalletManagerProvider'
+import {debugWalletInfo, features} from '../../kernel/features'
 import {confirmationMessages, errorMessages, txLabels} from '../../kernel/i18n/global-messages'
 import LocalizableError from '../../kernel/i18n/LocalizableError'
 import {isEmptyString} from '../../kernel/utils'
