+++ apps/wallet-mobile/src/components/ConfirmTxWithSpendingPasswordModal/ConfirmTxWithSpendingPasswordModal.tsx
@@ -3,8 +3,8 @@ import {App} from '@yoroi/types'
 import React, {useEffect, useRef, useState} from 'react'
 import {ActivityIndicator, StyleSheet, TextInput as RNTextInput, View} from 'react-native'
 
-import {debugWalletInfo, features} from '../../features'
 import {useSelectedWallet} from '../../features/WalletManager/common/hooks/useSelectedWallet'
+import {debugWalletInfo, features} from '../../kernel/features'
 import {useSignTxWithPassword, useSubmitTx} from '../../yoroi-wallets/hooks'
 import {YoroiSignedTx, YoroiUnsignedTx} from '../../yoroi-wallets/types'
 import {Button} from '../Button'
