+++ apps/wallet-mobile/src/features/Portfolio/useCases/PortfolioTokenDetails/PortfolioTokenDetailsScreen.tsx
@@ -5,8 +5,8 @@ import {Animated, NativeScrollEvent, NativeSyntheticEvent, StyleSheet} from 'rea
 import {Spacer} from '../../../../components'
 import {SafeArea} from '../../../../components/SafeArea'
 import {Tab, Tabs} from '../../../../components/Tabs'
+import {features} from '../../../../kernel/features'
 import {useMetrics} from '../../../../kernel/metrics/metricsManager'
-import {features} from '../../..'
 import {TxFilter} from '../../../Transactions/useCases/TxList/TxFilterProvider'
 import {TxList} from '../../../Transactions/useCases/TxList/TxList'
 import {usePortfolioTokenDetailContext} from '../../common/PortfolioTokenDetailContext'
