+++ apps/wallet-mobile/src/features/Portfolio/useCases/PortfolioTokenDetails/PortfolioTokenInfo/Overview/index.tsx
@@ -5,7 +5,7 @@ import React, {useState} from 'react'
 import {Linking, StyleSheet, Text, TouchableOpacity, View} from 'react-native'
 
 import {Accordion, CopyButton, Spacer} from '../../../../../../components'
-import {features} from '../../../../..'
+import {features} from '../../../../../../kernel/features'
 import {useSelectedWallet} from '../../../../../WalletManager/common/hooks/useSelectedWallet'
 import {TokenInfoIcon} from '../../../../common/TokenAmountItem/TokenInfoIcon'
 import {usePortfolioTokenDetailParams} from '../../../../common/useNavigateTo'
