+++ apps/wallet-mobile/src/legacy/Staking/DelegationConfirmation/DelegationConfirmation.tsx
@@ -8,8 +8,8 @@ import {useQueryClient} from 'react-query'
 import {KeyboardSpacer, Text, ValidatedTextInput} from '../../../components'
 import {ConfirmTx} from '../../../components/ConfirmTx'
 import {Space} from '../../../components/Space/Space'
-import {debugWalletInfo, features} from '../../../features'
 import {useSelectedWallet} from '../../../features/WalletManager/common/hooks/useSelectedWallet'
+import {debugWalletInfo, features} from '../../../kernel/features'
 import globalMessages, {txLabels} from '../../../kernel/i18n/global-messages'
 import {StakingCenterRoutes, useParams, useWalletNavigation} from '../../../kernel/navigation'
 import {NETWORKS} from '../../../yoroi-wallets/cardano/networks'
