+++ apps/wallet-mobile/src/legacy/Staking/PoolTransition/usePoolTransition.tsx
@@ -7,9 +7,9 @@ import * as React from 'react'
 import {defineMessages, useIntl} from 'react-intl'
 import {useQuery} from 'react-query'
 
-import {features} from '../../../features'
 import {useSelectedNetwork} from '../../../features/WalletManager/common/hooks/useSelectedNetwork'
 import {useSelectedWallet} from '../../../features/WalletManager/common/hooks/useSelectedWallet'
+import {features} from '../../../kernel/features'
 import {YoroiWallet} from '../../../yoroi-wallets/cardano/types'
 import {asQuantity, Quantities} from '../../../yoroi-wallets/utils'
 import {useStakingInfo} from '../../Dashboard/StakePoolInfos'
