+++ apps/wallet-mobile/src/legacy/Staking/StakingCenter/StakingCenter.tsx
@@ -8,10 +8,10 @@ import {View} from 'react-native'
 import {WebView, WebViewMessageEvent} from 'react-native-webview'
 
 import {PleaseWaitModal, Spacer} from '../../../components'
-import {features} from '../../../features'
 import {useSelectedWallet} from '../../../features/WalletManager/common/hooks/useSelectedWallet'
 import {showErrorDialog} from '../../../kernel/dialogs'
 import {isDev, isNightly} from '../../../kernel/env'
+import {features} from '../../../kernel/features'
 import {useLanguage} from '../../../kernel/i18n'
 import globalMessages from '../../../kernel/i18n/global-messages'
 import {logger} from '../../../kernel/logger/logger'
