+++ apps/wallet-mobile/src/yoroi-wallets/cardano/nfts.ts
@@ -3,7 +3,7 @@ import {domainNormalizer} from '@yoroi/resolver'
 import {Balance} from '@yoroi/types'
 import {z} from 'zod'
 
-import {features} from '../../features'
+import {features} from '../../kernel/features'
 import {getAssetFingerprint} from '../utils/format'
 import {toDisplayAssetName} from './api/utils'
 export const convertNft = (options: {
