+++ apps/wallet-mobile/translations/messages/src/features/SetupWallet/legacy/WalletForm.json
@@ -6,12 +6,12 @@
     "start": {
       "line": 145,
       "column": 24,
-      "index": 5375
+      "index": 5394
     },
     "end": {
       "line": 148,
       "column": 3,
-      "index": 5483
+      "index": 5502
     }
   },
   {
@@ -21,12 +21,12 @@
     "start": {
       "line": 149,
       "column": 20,
-      "index": 5505
+      "index": 5524
     },
     "end": {
       "line": 152,
       "column": 3,
-      "index": 5615
+      "index": 5634
     }
   },
   {
@@ -36,12 +36,12 @@
     "start": {
       "line": 153,
       "column": 18,
-      "index": 5635
+      "index": 5654
     },
     "end": {
       "line": 156,
       "column": 3,
-      "index": 5734
+      "index": 5753
     }
   },
   {
@@ -51,12 +51,12 @@
     "start": {
       "line": 157,
       "column": 31,
-      "index": 5767
+      "index": 5786
     },
     "end": {
       "line": 160,
       "column": 3,
-      "index": 5908
+      "index": 5927
     }
   },
   {
@@ -66,12 +66,12 @@
     "start": {
       "line": 161,
       "column": 28,
-      "index": 5938
+      "index": 5957
     },
     "end": {
       "line": 164,
       "column": 3,
-      "index": 6054
+      "index": 6073
     }
   },
   {
@@ -81,12 +81,12 @@
     "start": {
       "line": 165,
       "column": 28,
-      "index": 6084
+      "index": 6103
     },
     "end": {
       "line": 168,
       "column": 3,
-      "index": 6207
+      "index": 6226
     }
   }
 ]
