+++ apps/wallet-mobile/translations/messages/src/legacy/Staking/PoolTransition/usePoolTransition.json
@@ -6,12 +6,12 @@
     "start": {
       "line": 114,
       "column": 9,
-      "index": 4044
+      "index": 4051
     },
     "end": {
       "line": 117,
       "column": 3,
-      "index": 4142
+      "index": 4149
     }
   },
   {
@@ -21,12 +21,12 @@
     "start": {
       "line": 118,
       "column": 11,
-      "index": 4155
+      "index": 4162
     },
     "end": {
       "line": 122,
       "column": 3,
-      "index": 4351
+      "index": 4358
     }
   },
   {
@@ -36,12 +36,12 @@
     "start": {
       "line": 123,
       "column": 16,
-      "index": 4369
+      "index": 4376
     },
     "end": {
       "line": 127,
       "column": 3,
-      "index": 4575
+      "index": 4582
     }
   },
   {
@@ -51,12 +51,12 @@
     "start": {
       "line": 128,
       "column": 15,
-      "index": 4592
+      "index": 4599
     },
     "end": {
       "line": 131,
       "column": 3,
-      "index": 4685
+      "index": 4692
     }
   },
   {
@@ -66,12 +66,12 @@
     "start": {
       "line": 132,
       "column": 11,
-      "index": 4698
+      "index": 4705
     },
     "end": {
       "line": 135,
       "column": 3,
-      "index": 4783
+      "index": 4790
     }
   },
   {
@@ -81,12 +81,12 @@
     "start": {
       "line": 136,
       "column": 16,
-      "index": 4801
+      "index": 4808
     },
     "end": {
       "line": 139,
       "column": 3,
-      "index": 4896
+      "index": 4903
     }
   },
   {
@@ -96,12 +96,12 @@
     "start": {
       "line": 140,
       "column": 7,
-      "index": 4905
+      "index": 4912
     },
     "end": {
       "line": 143,
       "column": 3,
-      "index": 4981
+      "index": 4988
     }
   },
   {
@@ -111,12 +111,12 @@
     "start": {
       "line": 144,
       "column": 24,
-      "index": 5007
+      "index": 5014
     },
     "end": {
       "line": 147,
       "column": 3,
-      "index": 5144
+      "index": 5151
     }
   },
   {
@@ -126,12 +126,12 @@
     "start": {
       "line": 148,
       "column": 17,
-      "index": 5163
+      "index": 5170
     },
     "end": {
       "line": 151,
       "column": 3,
-      "index": 5297
+      "index": 5304
     }
   },
   {
@@ -141,12 +141,12 @@
     "start": {
       "line": 152,
       "column": 23,
-      "index": 5322
+      "index": 5329
     },
     "end": {
       "line": 155,
       "column": 3,
-      "index": 5452
+      "index": 5459
     }
   },
   {
@@ -156,12 +156,12 @@
     "start": {
       "line": 156,
       "column": 17,
-      "index": 5471
+      "index": 5478
     },
     "end": {
       "line": 159,
       "column": 3,
-      "index": 5585
+      "index": 5592
     }
   },
   {
@@ -171,12 +171,12 @@
     "start": {
       "line": 160,
       "column": 21,
-      "index": 5608
+      "index": 5615
     },
     "end": {
       "line": 163,
       "column": 3,
-      "index": 5722
+      "index": 5729
     }
   },
   {
@@ -186,12 +186,12 @@
     "start": {
       "line": 164,
       "column": 10,
-      "index": 5734
+      "index": 5741
     },
     "end": {
       "line": 167,
       "column": 3,
-      "index": 5821
+      "index": 5828
     }
   }
 ]
