+++ apps/wallet-mobile/translations/messages/src/legacy/Staking/StakingCenter/StakingCenter.json
@@ -6,12 +6,12 @@
     "start": {
       "line": 109,
       "column": 9,
-      "index": 3919
+      "index": 3926
     },
     "end": {
       "line": 112,
       "column": 3,
-      "index": 4027
+      "index": 4034
     }
   },
   {
@@ -21,12 +21,12 @@
     "start": {
       "line": 113,
       "column": 11,
-      "index": 4040
+      "index": 4047
     },
     "end": {
       "line": 116,
       "column": 3,
-      "index": 4206
+      "index": 4213
     }
   }
 ]
