+++ apps/wallet-mobile/translations/messages/src/yoroi-wallets/utils/format.json
@@ -4,14 +4,25 @@
     "defaultMessage": "!!!Today",
     "file": "src/yoroi-wallets/utils/format.ts",
     "start": {
-      "line": 238,
+<<<<<<< Updated upstream
+      "line": 228,
       "column": 9,
       "index": 7585
     },
     "end": {
       "line": 241,
       "column": 3,
-      "index": 7652
+      "index": 7351
+=======
+      "line": 237,
+      "column": 9,
+      "index": 7555
+    },
+    "end": {
+      "line": 240,
+      "column": 3,
+      "index": 7622
+>>>>>>> Stashed changes
     }
   },
   {
@@ -19,14 +30,25 @@
     "defaultMessage": "!!!Yesterday",
     "file": "src/yoroi-wallets/utils/format.ts",
     "start": {
-      "line": 242,
+<<<<<<< Updated upstream
+      "line": 232,
       "column": 13,
       "index": 7667
     },
     "end": {
       "line": 245,
       "column": 3,
-      "index": 7742
+      "index": 7441
+=======
+      "line": 241,
+      "column": 13,
+      "index": 7637
+    },
+    "end": {
+      "line": 244,
+      "column": 3,
+      "index": 7712
+>>>>>>> Stashed changes
     }
   },
   {
@@ -34,14 +56,25 @@
     "defaultMessage": "!!![Unknown asset name]",
     "file": "src/yoroi-wallets/utils/format.ts",
     "start": {
-      "line": 246,
+<<<<<<< Updated upstream
+      "line": 236,
       "column": 20,
       "index": 7764
     },
     "end": {
       "line": 249,
       "column": 3,
-      "index": 7857
+      "index": 7556
+=======
+      "line": 245,
+      "column": 20,
+      "index": 7734
+    },
+    "end": {
+      "line": 248,
+      "column": 3,
+      "index": 7827
+>>>>>>> Stashed changes
     }
   }
 ]
